· Metode per anul·lar venda modificat per evitar possibles errors

+++ src/main/java/org/udg/pds/cheapy/service/ProducteService.java
@@ -300,12 +300,14 @@ public class ProducteService
 
         Transaccio t = p.getTransaccio();
 
-        Valoracio vComprador = t.getValoracioComprador();
-        Valoracio vVenedor = t.getValoracioVenedor();
-
-        em.remove(vComprador);
-        em.remove(vVenedor);
-        em.remove(t); // eliminem la trasacció del producte
+        if(t != null) { // si hi ha transacció llavors obtenim les valoracions i la borrem
+            
+            Valoracio vComprador = t.getValoracioComprador();
+            Valoracio vVenedor = t.getValoracioVenedor();
+            if(vComprador != null) em.remove(vComprador);
+            if(vVenedor != null) em.remove(vVenedor);
+            em.remove(t);
+        }
 
         return new RESTService.ID(idProducte);
     }
